Move more XContent parser code that is test-only into test module
Just another round of #104261
