+++ server/src/main/java/org/elasticsearch/action/ingest/SimulatePipelineResponse.java
@@ -8,79 +8,21 @@
 
 package org.elasticsearch.action.ingest;
 
-import org.elasticsearch.ElasticsearchException;
 import org.elasticsearch.action.ActionResponse;
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
-import org.elasticsearch.xcontent.ConstructingObjectParser;
-import org.elasticsearch.xcontent.ParseField;
 import org.elasticsearch.xcontent.ToXContentObject;
 import org.elasticsearch.xcontent.XContentBuilder;
-import org.elasticsearch.xcontent.XContentParser;
-import org.elasticsearch.xcontent.XContentParser.Token;
 
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-import static org.elasticsearch.common.xcontent.XContentParserUtils.ensureExpectedToken;
-import static org.elasticsearch.xcontent.ConstructingObjectParser.constructorArg;
-
 public class SimulatePipelineResponse extends ActionResponse implements ToXContentObject {
     private String pipelineId;
     private boolean verbose;
     private List<SimulateDocumentResult> results;
-
-    @SuppressWarnings("unchecked")
-    public static final ConstructingObjectParser<SimulatePipelineResponse, Void> PARSER = new ConstructingObjectParser<>(
-        "simulate_pipeline_response",
-        true,
-        a -> {
-            List<SimulateDocumentResult> results = (List<SimulateDocumentResult>) a[0];
-            boolean verbose = false;
-            if (results.size() > 0) {
-                if (results.get(0) instanceof SimulateDocumentVerboseResult) {
-                    verbose = true;
-                }
-            }
-            return new SimulatePipelineResponse(null, verbose, results);
-        }
-    );
-    static {
-        PARSER.declareObjectArray(constructorArg(), (parser, context) -> {
-            Token token = parser.currentToken();
-            ensureExpectedToken(Token.START_OBJECT, token, parser);
-            SimulateDocumentResult result = null;
-            while ((token = parser.nextToken()) != Token.END_OBJECT) {
-                ensureExpectedToken(Token.FIELD_NAME, token, parser);
-                String fieldName = parser.currentName();
-                token = parser.nextToken();
-                if (token == Token.START_ARRAY) {
-                    if (fieldName.equals(SimulateDocumentVerboseResult.PROCESSOR_RESULT_FIELD)) {
-                        List<SimulateProcessorResult> results = new ArrayList<>();
-                        while ((token = parser.nextToken()) == Token.START_OBJECT) {
-                            results.add(SimulateProcessorResult.fromXContent(parser));
-                        }
-                        ensureExpectedToken(Token.END_ARRAY, token, parser);
-                        result = new SimulateDocumentVerboseResult(results);
-                    } else {
-                        parser.skipChildren();
-                    }
-                } else if (token.equals(Token.START_OBJECT)) {
-                    switch (fieldName) {
-                        case WriteableIngestDocument.DOC_FIELD -> result = new SimulateDocumentBaseResult(
-                            WriteableIngestDocument.INGEST_DOC_PARSER.apply(parser, null).getIngestDocument()
-                        );
-                        case "error" -> result = new SimulateDocumentBaseResult(ElasticsearchException.fromXContent(parser));
-                        default -> parser.skipChildren();
-                    }
-                } // else it is a value skip it
-            }
-            assert result != null;
-            return result;
-        }, new ParseField(Fields.DOCUMENTS));
-    }
 
     public SimulatePipelineResponse(StreamInput in) throws IOException {
         super(in);
@@ -134,10 +76,6 @@ public class SimulatePipelineResponse extends ActionResponse implements ToXConte
         builder.endArray();
         builder.endObject();
         return builder;
-    }
-
-    public static SimulatePipelineResponse fromXContent(XContentParser parser) {
-        return PARSER.apply(parser, null);
     }
 
     static final class Fields {
