+++ server/src/main/java/org/elasticsearch/index/reindex/BulkByScrollResponse.java
@@ -8,23 +8,14 @@
 
 package org.elasticsearch.index.reindex;
 
-import org.elasticsearch.ElasticsearchException;
-import org.elasticsearch.ElasticsearchParseException;
 import org.elasticsearch.action.ActionResponse;
 import org.elasticsearch.action.bulk.BulkItemResponse.Failure;
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
 import org.elasticsearch.core.Nullable;
 import org.elasticsearch.core.TimeValue;
-import org.elasticsearch.index.reindex.BulkByScrollTask.Status;
-import org.elasticsearch.index.reindex.ScrollableHitSource.SearchFailure;
-import org.elasticsearch.rest.RestStatus;
-import org.elasticsearch.xcontent.ObjectParser;
-import org.elasticsearch.xcontent.ParseField;
 import org.elasticsearch.xcontent.ToXContentFragment;
 import org.elasticsearch.xcontent.XContentBuilder;
-import org.elasticsearch.xcontent.XContentParser;
-import org.elasticsearch.xcontent.XContentParser.Token;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -33,7 +24,6 @@ import java.util.List;
 import static java.lang.Math.max;
 import static java.lang.Math.min;
 import static java.util.Objects.requireNonNull;
-import static org.elasticsearch.common.xcontent.XContentParserUtils.ensureExpectedToken;
 import static org.elasticsearch.core.TimeValue.timeValueNanos;
 
 /**
@@ -46,23 +36,9 @@ public class BulkByScrollResponse extends ActionResponse implements ToXContentFr
     private List<ScrollableHitSource.SearchFailure> searchFailures;
     private boolean timedOut;
 
-    private static final String TOOK_FIELD = "took";
-    private static final String TIMED_OUT_FIELD = "timed_out";
-    private static final String FAILURES_FIELD = "failures";
-
-    @SuppressWarnings("unchecked")
-    private static final ObjectParser<BulkByScrollResponseBuilder, Void> PARSER = new ObjectParser<>(
-        "bulk_by_scroll_response",
-        true,
-        BulkByScrollResponseBuilder::new
-    );
-    static {
-        PARSER.declareLong(BulkByScrollResponseBuilder::setTook, new ParseField(TOOK_FIELD));
-        PARSER.declareBoolean(BulkByScrollResponseBuilder::setTimedOut, new ParseField(TIMED_OUT_FIELD));
-        PARSER.declareObjectArray(BulkByScrollResponseBuilder::setFailures, (p, c) -> parseFailure(p), new ParseField(FAILURES_FIELD));
-        // since the result of BulkByScrollResponse.Status are mixed we also parse that in this
-        Status.declareFields(PARSER);
-    }
+    static final String TOOK_FIELD = "took";
+    static final String TIMED_OUT_FIELD = "timed_out";
+    static final String FAILURES_FIELD = "failures";
 
     public BulkByScrollResponse(StreamInput in) throws IOException {
         super(in);
@@ -195,7 +171,7 @@ public class BulkByScrollResponse extends ActionResponse implements ToXContentFr
         builder.field(TOOK_FIELD, took.millis());
         builder.field(TIMED_OUT_FIELD, timedOut);
         status.innerXContent(builder, params);
-        builder.startArray("failures");
+        builder.startArray(FAILURES_FIELD);
         for (Failure failure : bulkFailures) {
             builder.startObject();
             failure.toXContent(builder, params);
@@ -206,59 +182,6 @@ public class BulkByScrollResponse extends ActionResponse implements ToXContentFr
         }
         builder.endArray();
         return builder;
-    }
-
-    public static BulkByScrollResponse fromXContent(XContentParser parser) {
-        return PARSER.apply(parser, null).buildResponse();
-    }
-
-    private static Object parseFailure(XContentParser parser) throws IOException {
-        ensureExpectedToken(Token.START_OBJECT, parser.currentToken(), parser);
-        Token token;
-        String index = null;
-        String id = null;
-        Integer status = null;
-        Integer shardId = null;
-        String nodeId = null;
-        ElasticsearchException bulkExc = null;
-        ElasticsearchException searchExc = null;
-        while ((token = parser.nextToken()) != Token.END_OBJECT) {
-            ensureExpectedToken(Token.FIELD_NAME, token, parser);
-            String name = parser.currentName();
-            token = parser.nextToken();
-            if (token == Token.START_ARRAY) {
-                parser.skipChildren();
-            } else if (token == Token.START_OBJECT) {
-                switch (name) {
-                    case SearchFailure.REASON_FIELD -> searchExc = ElasticsearchException.fromXContent(parser);
-                    case Failure.CAUSE_FIELD -> bulkExc = ElasticsearchException.fromXContent(parser);
-                    default -> parser.skipChildren();
-                }
-            } else if (token == Token.VALUE_STRING) {
-                switch (name) {
-                    // This field is the same as SearchFailure.index
-                    case Failure.INDEX_FIELD -> index = parser.text();
-                    case Failure.ID_FIELD -> id = parser.text();
-                    case SearchFailure.NODE_FIELD -> nodeId = parser.text();
-                }
-            } else if (token == Token.VALUE_NUMBER) {
-                switch (name) {
-                    case Failure.STATUS_FIELD -> status = parser.intValue();
-                    case SearchFailure.SHARD_FIELD -> shardId = parser.intValue();
-                }
-            }
-        }
-        if (bulkExc != null) {
-            return new Failure(index, id, bulkExc, RestStatus.fromCode(status));
-        } else if (searchExc != null) {
-            if (status == null) {
-                return new SearchFailure(searchExc, index, shardId, nodeId);
-            } else {
-                return new SearchFailure(searchExc, index, shardId, nodeId, RestStatus.fromCode(status));
-            }
-        } else {
-            throw new ElasticsearchParseException("failed to parse failures array. At least one of {reason,cause} must be present");
-        }
     }
 
     @Override
