+++ server/src/test/java/org/elasticsearch/action/admin/cluster/storedscripts/GetStoredScriptResponseTests.java
@@ -12,6 +12,8 @@ import org.elasticsearch.common.io.stream.Writeable;
 import org.elasticsearch.script.Script;
 import org.elasticsearch.script.StoredScriptSource;
 import org.elasticsearch.test.AbstractXContentSerializingTestCase;
+import org.elasticsearch.xcontent.ConstructingObjectParser;
+import org.elasticsearch.xcontent.ObjectParser;
 import org.elasticsearch.xcontent.XContentParser;
 import org.elasticsearch.xcontent.XContentType;
 
@@ -20,11 +22,41 @@ import java.util.Collections;
 import java.util.Map;
 import java.util.function.Predicate;
 
+import static org.elasticsearch.xcontent.ConstructingObjectParser.constructorArg;
+import static org.elasticsearch.xcontent.ConstructingObjectParser.optionalConstructorArg;
+
 public class GetStoredScriptResponseTests extends AbstractXContentSerializingTestCase<GetStoredScriptResponse> {
+
+    private static final ConstructingObjectParser<GetStoredScriptResponse, String> PARSER = new ConstructingObjectParser<>(
+        "GetStoredScriptResponse",
+        true,
+        (a, c) -> {
+            String id = (String) a[0];
+            boolean found = (Boolean) a[1];
+            StoredScriptSource scriptSource = (StoredScriptSource) a[2];
+            return found ? new GetStoredScriptResponse(id, scriptSource) : new GetStoredScriptResponse(id, null);
+        }
+    );
+
+    static {
+        PARSER.declareField(constructorArg(), (p, c) -> p.text(), GetStoredScriptResponse._ID_PARSE_FIELD, ObjectParser.ValueType.STRING);
+        PARSER.declareField(
+            constructorArg(),
+            (p, c) -> p.booleanValue(),
+            GetStoredScriptResponse.FOUND_PARSE_FIELD,
+            ObjectParser.ValueType.BOOLEAN
+        );
+        PARSER.declareField(
+            optionalConstructorArg(),
+            (p, c) -> StoredScriptSource.fromXContent(p, true),
+            GetStoredScriptResponse.SCRIPT,
+            ObjectParser.ValueType.OBJECT
+        );
+    }
 
     @Override
     protected GetStoredScriptResponse doParseInstance(XContentParser parser) throws IOException {
-        return GetStoredScriptResponse.fromXContent(parser);
+        return PARSER.apply(parser, null);
     }
 
     @Override
