+++ server/src/test/java/org/elasticsearch/action/get/MultiGetResponseTests.java
@@ -7,21 +7,33 @@
  */
 package org.elasticsearch.action.get;
 
+import org.elasticsearch.ElasticsearchException;
 import org.elasticsearch.common.bytes.BytesReference;
+import org.elasticsearch.common.logging.DeprecationLogger;
 import org.elasticsearch.index.get.GetResult;
+import org.elasticsearch.rest.action.document.RestMultiGetAction;
 import org.elasticsearch.test.ESTestCase;
+import org.elasticsearch.xcontent.ParseField;
 import org.elasticsearch.xcontent.ToXContent;
 import org.elasticsearch.xcontent.XContentFactory;
 import org.elasticsearch.xcontent.XContentParser;
 import org.elasticsearch.xcontent.XContentType;
 
 import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
 
 import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
 
 public class MultiGetResponseTests extends ESTestCase {
+
+    private static final DeprecationLogger deprecationLogger = DeprecationLogger.getLogger(MultiGetResponse.class);
+
+    private static final ParseField TYPE = new ParseField("_type");
+
+    private static final ParseField ERROR = new ParseField("error");
 
     public void testFromXContent() throws IOException {
         for (int runs = 0; runs < 20; runs++) {
@@ -30,7 +42,7 @@ public class MultiGetResponseTests extends ESTestCase {
             BytesReference shuffled = toShuffledXContent(expected, xContentType, ToXContent.EMPTY_PARAMS, false);
             MultiGetResponse parsed;
             try (XContentParser parser = createParser(XContentFactory.xContent(xContentType), shuffled)) {
-                parsed = MultiGetResponse.fromXContent(parser);
+                parsed = parseInstance(parser);
                 assertNull(parser.nextToken());
             }
             assertNotSame(expected, parsed);
@@ -75,6 +87,80 @@ public class MultiGetResponseTests extends ESTestCase {
             }
         }
         return new MultiGetResponse(items);
+    }
+
+    public static MultiGetResponse parseInstance(XContentParser parser) throws IOException {
+        String currentFieldName = null;
+        List<MultiGetItemResponse> items = new ArrayList<>();
+        for (XContentParser.Token token = parser.nextToken(); token != XContentParser.Token.END_OBJECT; token = parser.nextToken()) {
+            switch (token) {
+                case FIELD_NAME:
+                    currentFieldName = parser.currentName();
+                    break;
+                case START_ARRAY:
+                    if (MultiGetResponse.DOCS.getPreferredName().equals(currentFieldName)) {
+                        for (token = parser.nextToken(); token != XContentParser.Token.END_ARRAY; token = parser.nextToken()) {
+                            if (token == XContentParser.Token.START_OBJECT) {
+                                items.add(parseItem(parser));
+                            }
+                        }
+                    }
+                    break;
+                default:
+                    // If unknown tokens are encounter then these should be ignored, because
+                    // this is parsing logic on the client side.
+                    break;
+            }
+        }
+        return new MultiGetResponse(items.toArray(new MultiGetItemResponse[0]));
+    }
+
+    private static MultiGetItemResponse parseItem(XContentParser parser) throws IOException {
+        String currentFieldName = null;
+        String index = null;
+        String id = null;
+        ElasticsearchException exception = null;
+        GetResult getResult = null;
+        for (XContentParser.Token token = parser.nextToken(); token != XContentParser.Token.END_OBJECT; token = parser.nextToken()) {
+            switch (token) {
+                case FIELD_NAME:
+                    currentFieldName = parser.currentName();
+                    if (MultiGetResponse.INDEX.match(currentFieldName, parser.getDeprecationHandler()) == false
+                        && MultiGetResponse.ID.match(currentFieldName, parser.getDeprecationHandler()) == false
+                        && ERROR.match(currentFieldName, parser.getDeprecationHandler()) == false) {
+                        getResult = GetResult.fromXContentEmbedded(parser, index, id);
+                    }
+                    break;
+                case VALUE_STRING:
+                    if (MultiGetResponse.INDEX.match(currentFieldName, parser.getDeprecationHandler())) {
+                        index = parser.text();
+                    } else if (TYPE.match(currentFieldName, parser.getDeprecationHandler())) {
+                        deprecationLogger.compatibleCritical("mget_with_types", RestMultiGetAction.TYPES_DEPRECATION_MESSAGE);
+                    } else if (MultiGetResponse.ID.match(currentFieldName, parser.getDeprecationHandler())) {
+                        id = parser.text();
+                    }
+                    break;
+                case START_OBJECT:
+                    if (ERROR.match(currentFieldName, parser.getDeprecationHandler())) {
+                        exception = ElasticsearchException.fromXContent(parser);
+                    }
+                    break;
+                default:
+                    // If unknown tokens are encounter then these should be ignored, because
+                    // this is parsing logic on the client side.
+                    break;
+            }
+            if (getResult != null) {
+                break;
+            }
+        }
+
+        if (exception != null) {
+            return new MultiGetItemResponse(null, new MultiGetResponse.Failure(index, id, exception));
+        } else {
+            GetResponse getResponse = new GetResponse(getResult);
+            return new MultiGetItemResponse(getResponse, null);
+        }
     }
 
 }
